diagd: Move estats update to its own thread
The estat update thread was just posting a message to the
AmbassadorEventWatcher thread.
Then this AmbassadorEventWatcher thread was processing sequentially all
the message.
This thread is also updating the config. This message can take some
time, more than 20 seconds.
20 seconds is the delay used by ambassador to declare envoy dead when it
could not get an update status from envoy.
But when the config takes more than 20 seconds, this blocks the message
doing this envoy update status, and then ambassador alive endpoint is
starting to return 503.
Then, at the k8s level, the liveness probe is failing and the container
is restarted.

In order to prevent this lag in the envoy update state, the timer which
was posting the update message is now directly doing the update itself.